requirements.txt
Django>=3.2,<3.3
djangorestframework>=3.12.4,<3.13
djangorestframework-simplejwt>=4.6.0,<4.7.0
drf-extra-fields>=3.1.1,<3.2
psycopg2>=2.8.6,<2.9
gunicorn>=20.1.0,<20.2.0